+++ tests/tests_grasp.py
@@ -1,7 +1,10 @@
 import unittest
+import sys
 
 from networkx import DiGraph
 from numpy import array
+
+sys.path.append("../")
 
 from cspy.algorithms.grasp import GRASP
 
@@ -11,6 +14,7 @@ class TestsGRASP(unittest.TestCase):
     Tests for finding the resource constrained shortest
     path of simple DiGraph using the GRASP algorithm.
     """
+
     def setUp(self):
         self.max_res, self.min_res = [5, 5], [0, 0]